+++ src/app/backend/service/UserService.ts
@@ -1,13 +1,13 @@
 import { User } from "@model/user";
-import { dbConnect } from '@lib/dbConnect'
-import { UserRepository } from '@repository/UserRepository';
+import { dbConnect } from "@lib/dbConnect";
+import { UserRepository } from "@repository/UserRepository";
 
 export async function createUser(user:User) : Promise<User>  {
     await dbConnect();
     return await UserRepository.create(user);
 }
 
-export async function getAllUsers() : Promise<Array<User>> {
+export async function getAllUsers() : Promise<User[]> {
     await dbConnect();
     return await UserRepository.find();
 }
