+++ src/app/usecase/user.ts
@@ -1,0 +1,17 @@
+import { User } from "@model/user";
+
+export async function getAllUsers() : Promise<User[]> {
+    try {
+        const res = await fetch("http://localhost:3000/api/users", {
+          cache: "no-store",
+        });
+    
+        if (!res.ok) {
+          throw new Error("Failed to fetch users");
+        }
+    
+        return res.json();
+      } catch (error) {
+        console.log("Error loading users: ", error);
+      }
+}
